GridmapVisualisation Update.
Fixes a minor bug with the heightfield.
(The Heightfield is actually drawn from cell center to
cell center so for an accurate visualization of the data
it has to be translated by half a cell on the x and y axes
and the applied texture has to be rescaled.)

Added a Q_PROPERTY to en-/disable the use of NPOT textures.

Added a Q_PROPERTY to en-/disable the heightmap.
If disabled a plane is created instead.

Adds a Q_PROPERTY to en-/disable interpolation
of colors between cells.